fix(reg): widen field_relax's `stream` to intptr_t
`field_relax` was the single binding in ext.cpp still declaring
`int stream`; the other 47 (including its own sibling `field_relax_rls`
and `flow_relax`) already take `intptr_t`. It was added in #21, in
parallel with the widening pass in #22, and never converted.

nanobind's `int` caster range-checks rather than truncating, so any
genuine 64-bit CUDA stream handle -- what both Python callers pass,
`torch.cuda.current_stream(...).cuda_stream` and
`cupy.cuda.get_current_stream().ptr` -- was rejected with TypeError.
`fastfields.torch.field_relax` / `fastfields.cupy.field_relax` therefore
raised on essentially every real GPU call; latent only because there is
no GPU in CI.

Adds a regression test that calls every field_*/flow_* binding with a
1<<40 stream handle, plus a completeness guard so a new binding has to
be added to the table.

Refs fastfields/fastfields-lib#69 (item 1)

+++ src/ext.cpp
@@ -455,7 +455,7 @@ NB_MODULE(_core, m) {
            std::optional<std::vector<double>> absolute,
            std::optional<std::vector<double>> membrane,
            std::optional<std::vector<double>> bending, int8_t bound, int ndim,
-           int nb_iter, int stream) {
+           int nb_iter, intptr_t stream) {
             DLTensor s = to_dltensor(sol), h = to_dltensor(hes),
                      g = to_dltensor(grd);
             ff::field_relax(s, h, g, vec_ptr(voxel_size), vec_ptr(absolute),
